Split python event store tests to easier cover SIO and ROOT
